# Execution refactoring and performance improvements