… Cargo.lock
… Cargo.toml
… crates/contracts/core/ab-contracts-executor/Cargo.toml
… crates/contracts/core/ab-contracts-executor/src/slots.rs
… crates/contracts/core/ab-contracts-io-type/src/lib.rs
… crates/contracts/core/ab-contracts-io-type/src/maybe_data.rs
… crates/contracts/core/ab-contracts-io-type/src/metadata.rs
… crates/contracts/core/ab-contracts-io-type/src/metadata/type_details.rs
… crates/contracts/core/ab-contracts-io-type/src/trivial_type.rs
… crates/contracts/core/ab-contracts-io-type/src/variable_bytes.rs
… crates/contracts/core/ab-contracts-io-type/src/variable_elements.rs
… crates/contracts/core/ab-contracts-macros-impl/src/contract/method.rs
… crates/contracts/core/ab-contracts-trivial-type-derive/src/lib.rs
… crates/contracts/example/ab-example-contract-flipper/Cargo.toml
… crates/contracts/example/ab-example-contract-flipper/benches/flip.rs
… crates/contracts/example/ab-example-contract-wallet/Cargo.toml
… crates/contracts/example/ab-example-contract-wallet/benches/example-contract-wallet.rs
… crates/contracts/system/ab-system-contract-simple-wallet-base/src/lib.rs
… crates/contracts/system/ab-system-contract-simple-wallet-base/src/payload.rs