After thinking about for a while and estimating the complexity, I decided to drop support for calling multiple methods. Simple VM instantiation will likely take non-negligible amount of time and general number of transactions should be sufficient to utilize multiple CPU cores without allowing a single transaction to use more than one core.

Now all locking is gone from native execution environment as well as some heap allocations.

There are also additional benchmark on flipper that measures performance of simple method calls and transaction emulation.

As the result simple flipper call performance more than doubled from ~5M flips/s to over 12M flips/s:
```
flipper/direct          time:   [82.054 ns 82.290 ns 82.568 ns]
                        thrpt:  [12.111 Melem/s 12.152 Melem/s 12.187 Melem/s]
flipper/transaction     time:   [97.516 ns 97.757 ns 98.039 ns]
                        thrpt:  [10.200 Melem/s 10.229 Melem/s 10.255 Melem/s]
```

Smart contract calls have also improved, I decided to offer a separate method for doing execution of pre-verified transactions since verification is the majority of the transaction processing cost right now:
```
example-wallet/verify-only
                        time:   [18.312 µs 18.335 µs 18.361 µs]
                        thrpt:  [54.464 Kelem/s 54.539 Kelem/s 54.608 Kelem/s]
example-wallet/execute-only
                        time:   [3.1777 µs 3.1829 µs 3.1888 µs]
                        thrpt:  [313.60 Kelem/s 314.18 Kelem/s 314.70 Kelem/s]
```

For simple transactions over 300k transactions/s can be processed, while verification is much slower.

There is no `unsafe` in slots anymore, which is a relief, but there is a bit more `unsafe` in FFI call module due to tricky referencing of slots and environment. I think some of that might be simplified eventually, but it is what it is for now. At least Miri seems to be happy so far and I'll add `loom` at some point too to cover more ground.